Removed dependency on nnUNetDataset, moved newDataset to diff file, fixed target shape issue
